Moving files into CSF/extensions
