+++ packages/lib/src/components/internal/SecuredFields/lib/CSF/extensions/handleBinValue.ts
