+++ packages/lib/src/components/internal/SecuredFields/lib/CSF/extensions/handleConfig.ts
@@ -1,4 +1,4 @@
-import { getCardGroupTypes } from '../../core/utils/getCardGroupTypes';
+import { getCardGroupTypes } from '../utils/getCardGroupTypes';
 import { NON_CREDIT_CARD_TYPE_SECURED_FIELDS, SF_VERSION } from '../../configuration/constants';
 import * as logger from '../../utilities/logger';
