+++ packages/lib/src/components/internal/SecuredFields/lib/CSF/extensions/handleFocus.ts
@@ -1,4 +1,4 @@
-import { SFFeedbackObj, CbObjOnFocus } from '../../../types';
+import { SFFeedbackObj, CbObjOnFocus } from '../../types';
 
 // Call focus callback and store which field currently has focus
 export function handleFocus(pFeedbackObj: SFFeedbackObj): void {
