+++ packages/lib/src/components/internal/SecuredFields/lib/CSF/extensions/handleValidation.ts
@@ -1,6 +1,6 @@
-import { makeCallbackObjectsValidation } from '../../core/utils/callbackUtils';
+import { makeCallbackObjectsValidation } from '../utils/callbackUtils';
 import { removeEncryptedElement } from '../../ui/encryptedElements';
-import { processErrors } from '../../core/utils/processErrors';
+import { processErrors } from '../utils/processErrors';
 import { existy } from '../../utilities/commonUtils';
 import { ENCRYPTED_SECURITY_CODE, ENCRYPTED_CARD_NUMBER } from '../../configuration/constants';
 import { SFFeedbackObj, CbObjOnFieldValid } from '../../types';
