+++ packages/lib/src/components/internal/SecuredFields/lib/CSF/utils/tabbing/tabScenarioCreditCard.ts
