+++ packages/lib/src/components/internal/SecuredFields/lib/securedField/utils/processPlaceholders.ts
@@ -1,11 +1,11 @@
-import { resolvePlaceholders } from '../../../../utils';
-import { SFPlaceholdersObject, SFInternalConfig } from '../../../securedField/AbstractSecuredField';
+import { resolvePlaceholders } from '../../../utils';
+import { SFPlaceholdersObject, SFInternalConfig } from '../AbstractSecuredField';
 import {
     ENCRYPTED_SECURITY_CODE,
     ENCRYPTED_SECURITY_CODE_3_DIGITS,
     ENCRYPTED_SECURITY_CODE_4_DIGITS,
     GIFT_CARD
-} from '../../../configuration/constants';
+} from '../../configuration/constants';
 
 /**
  * Create placeholders with a value from the relevant translation file
